+++ internal/component/prometheus/exporter/ssh/ssh_exporter/ssh_exporter.go
@@ -2,7 +2,6 @@ package ssh_exporter
 
 import (
 	"github.com/go-kit/log"
-	"github.com/go-kit/log/level"
 	"github.com/grafana/alloy/internal/static/integrations"
 	integrations_v2 "github.com/grafana/alloy/internal/static/integrations/v2"
 	"github.com/grafana/alloy/internal/static/integrations/v2/metricsutils"
@@ -18,12 +17,7 @@ func (c *Config) InstanceKey(agentKey string) (string, error) {
 }
 
 func (c *Config) NewIntegration(logger log.Logger) (integrations.Integration, error) {
-	// Adjust the logger based on VerboseLogging
-	if c.VerboseLogging {
-		logger = level.NewFilter(logger, level.AllowDebug())
-	} else {
-		logger = level.NewFilter(logger, level.AllowInfo())
-	}
+	// Use the provided logger (global filters control verbosity)
 
 	var collectors []prometheus.Collector
 
@@ -37,8 +31,7 @@ func (c *Config) NewIntegration(logger log.Logger) (integrations.Integration, er
 	}
 
 	return integrations.NewCollectorIntegration(
-		c.Name(),
-		integrations.WithCollectors(collectors...),
+		c.Name(), integrations.WithCollectors(collectors...),
 	), nil
 }